Revert type change of `hashlen` in `rsa.c`
The type of `hashlen` is not uniform across the library, but at least consistently
`unsigned int` within the RSA module. Until we can change the type to `size_t`
everywhere in the next API changing release, it's probably better to at least
keep it consistent within each module.

+++ library/rsa.c
@@ -1180,7 +1180,7 @@ int mbedtls_rsa_rsassa_pss_sign( mbedtls_rsa_context *ctx,
  *
  */
 static int rsa_rsassa_pkcs1_v15_encode( mbedtls_md_type_t md_alg,
-                                        size_t hashlen,
+                                        unsigned int hashlen,
                                         const unsigned char *hash,
                                         size_t sig_len,
                                         unsigned char *dst )
